Устранение замечаний от 29.03.22
Поправлены отступы, сделана структура с константами для текстов, ключей(вместо енама), тайтла таббара. Текствью привязана к низу вью, добавлена прокрутка. Исправлены методы, чтобы не перегружать вьюдидлоад.
